Updated class packages to http
