+++ appserver/ejb/ejb-http-remoting/client/pom.xml
@@ -43,7 +43,7 @@
     
     <parent>
         <groupId>fish.payara.ejb.http</groupId>
-        <artifactId>ejb-rest-remoting</artifactId>
+        <artifactId>ejb-http-remoting</artifactId>
         <version>5.192-SNAPSHOT</version>
     </parent>
