+++ appserver/ejb/ejb-http-remoting/client/src/main/java/fish/payara/ejb/http/client/EjbRestProxyHandler.java
@@ -1,4 +1,4 @@
-package fish.payara.ejb.rest.client;
+package fish.payara.ejb.http.client;
 
 import static javax.naming.Context.SECURITY_CREDENTIALS;
 import static javax.naming.Context.SECURITY_PRINCIPAL;
